+++ src/libraries/recoil/user.ts
@@ -5,7 +5,6 @@ import { atom } from "recoil";
 const userState = atom<User | null>({
   key: `userState${uuidv4()}`,
   default: { id: 0, name: "test", type: "staff" },
-
 });
 
 export default userState;
