+++ src/pages/sfaff/StaffNoticePage.tsx
@@ -1,30 +1,56 @@
 import { Box, Stack, Typography, styled } from "@mui/material";
-import { AdminTable } from "@components/admin";
-
+import { AdminNoticeCard } from "@components/admin";
+import CSearchBox from "@components/common/atom/C-SearchBox";
+import { Link } from "react-router-dom";
+import RoutePath from "@routes/routePath";
 import CButton from "@components/common/atom/C-Button";
 
 import PaginationComponent from "@components/common/pagination";
-import AdminNoticeWriteForm from "@components/admin/adminNotice/adminNoticeWriteForm";
+import AdminNoticeCardDeatil from "@components/admin/adminNotice/adminNoticeCardDetail";
 
-const AdminNoticeWritePage = () => {
+const mock = [1, 2, 3, 4, 5, 6, 7];
+
+const StaffNoticePage = () => {
   return (
     <Container>
       <div>
         <Title variant="h1">공지</Title>
         <AdminInoutSubTitleContainer>
-          <AdminInoutSubTitleLeftContainer></AdminInoutSubTitleLeftContainer>
+          <AdminInoutSubTitleLeftContainer>
+            <CSearchBox
+              value={""}
+              onChange={() => null}
+              placeholder={"검색 내용을 입력해주세요."}
+              borderColor="#8C8E94"
+            />
+          </AdminInoutSubTitleLeftContainer>
           <AdminInoutSubTitleRightContainer>
+            <ButtonListLayout>
+              <ButtonLayout width={"124px"}>
+                <CButton buttonType={"primaryBlack"}>내 환자 대상 공지</CButton>
+              </ButtonLayout>
+              <ButtonLayout width={"121px"}>
+                <CButton buttonType={"primaryBlack"}>내가 작성한 공지</CButton>
+              </ButtonLayout>
+              <ButtonLayout width={"150px"}>
+                <CButton buttonType={"primaryBlack"}>내 근무 시간동안 공지</CButton>
+              </ButtonLayout>
+            </ButtonListLayout>
             <ButtonLayout width={"148px"}>
-              <CButton buttonType={"primarySpaure"}>공지 작성</CButton>
+              <Link to={RoutePath.StaffNoticeWrite}>
+                <CButton buttonType={"primarySpaureWhite"}>공지 작성</CButton>
+              </Link>
             </ButtonLayout>
           </AdminInoutSubTitleRightContainer>
         </AdminInoutSubTitleContainer>
       </div>
       <TableLayout>
         <AdminNoticeListLayout>
-          <AdminTable />
+          {mock.map(item => {
+            return <AdminNoticeCard key={item} />;
+          })}
         </AdminNoticeListLayout>
-        <AdminNoticeWriteForm />
+        <AdminNoticeCardDeatil />
       </TableLayout>
       <FooterLayout>
         <PaginationComponent totalPage={5} />
@@ -33,7 +59,7 @@ const AdminNoticeWritePage = () => {
   );
 };
 
-export default AdminNoticeWritePage;
+export default StaffNoticePage;
 
 /** styles */
 
@@ -42,9 +68,11 @@ const Container = styled(Stack)({
 });
 
 const AdminInoutSubTitleContainer = styled(Box)({
+  width: "100%",
   marginTop: "17.66px",
   display: "flex",
   justifyContent: "space-between",
+  gap: "22px",
 });
 
 const AdminNoticeListLayout = styled(Box)({
@@ -54,7 +82,7 @@ const AdminNoticeListLayout = styled(Box)({
 const AdminInoutSubTitleLeftContainer = styled(Box)({
   display: "flex",
   width: "50%",
-  gap: "20px",
+  // gap: "20px",
   alignItems: "center",
 });
 const AdminInoutSubTitleRightContainer = styled(Box)({
@@ -63,10 +91,11 @@ const AdminInoutSubTitleRightContainer = styled(Box)({
   alignItems: "center",
   gap: "15.66px",
 
-  justifyContent: "end",
+  justifyContent: "space-between",
 });
 const ButtonLayout = styled(Box)(width => ({
   width: `${width}`,
+  maxHeight: "34px",
 }));
 const ButtonListLayout = styled(Box)({
   display: "flex",
